Rewritting from scratch.
Uses a threading system that takes each user's conversation into a
thread. When the bot thinks that the chat is finished, it sets itself to
closed and the server removes it from the active list. GC should remove
it at that point. Or shortly after at least.

Currently all that is working is "/torrent <SEARCH TERM>" which just
returns the names of the torrents on the first page of the kat.cr page